Convert from plugin to core feature
